rename one word files
