+++ src/view/filters-menu.js
